int to uint8_t

+++ nRF24-Nadajnik/src/nadajnik.h
@@ -53,8 +53,8 @@ class Nadajnik{
 
         struct WhistleData
         {
-            int     device = EWhistle;
-            int     command = EDefaultState;
+            uint8_t device = EWhistle;
+            uint8_t command = EDefaultState;
         };
         WhistleData whistleData;
